undo renaming main coordinator
