+++ WireUI/Sources/WireMainNavigation/Protocols/Containers/MainTabBarControllerProtocol.swift
@@ -26,7 +26,7 @@ import UIKit
 /// that a conforming tab bar controller must manage.
 
 @MainActor
-public protocol MainTabBarControllerProtocol: UITabBarController, MainContainerViewController {
+public protocol MainTabBarControllerProtocol: UITabBarController, MainContainerViewControllerProtocol {
 
     /// The selected content (tab).
     var selectedContent: MainTabBarControllerContent { get set }
