+++ wire-ios/Wire-iOS/Sources/UserInterface/MainController/MainCoordinator+SidebarViewControllerDelegate.swift
@@ -19,7 +19,7 @@
 import WireMainNavigation
 import WireSidebar
 
-extension GenericMainCoordinator: SidebarViewControllerDelegate where Dependencies.ConversationFilter == Wire.ConversationFilter {
+extension MainCoordinator: SidebarViewControllerDelegate where Dependencies.ConversationFilter == Wire.ConversationFilter {
 
     @MainActor
     public func sidebarViewControllerDidSelectAccountImage(_ viewController: SidebarViewController) {
