+++ wire-ios/Wire-iOS/Sources/UserInterface/MainController/MainCoordinator+StartUIDelegate.swift
@@ -19,7 +19,8 @@
 import WireDataModel
 import WireMainNavigation
 
-extension GenericMainCoordinator: StartUIDelegate where Dependencies.ConversationModel == ZMConversation, Dependencies.User == any UserType {
+// TODO: rename?
+extension MainCoordinator: StartUIDelegate where Dependencies.ConversationModel == ZMConversation, Dependencies.User == any UserType {
 
     @MainActor
     func startUIViewController(_ viewController: StartUIViewController, didSelect user: any UserType) {
